Unify authentication and profile logic
