+++ login.js
@@ -33,16 +33,73 @@ import {
   firebaseConfig,
   setPassphrase,
   getPassphrase,
-  clearPassphrase,
 } from './firebase-config.js';
 import { encryptString, decryptString } from './crypto.js';
 import { fetchResponsavelFinanceiroUsuarios } from './responsavel-financeiro.js';
 import { showToast } from './utils.js';
-import { loadUserProfile } from './user-profile.js';
 
 const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
 const auth = getAuth(app);
 const db = getFirestore(app);
+
+const profileCache = {};
+
+export async function loadUserProfile(uid) {
+  if (!uid) return null;
+  if (profileCache[uid]) return profileCache[uid];
+  const key = `userProfile:${uid}`;
+  try {
+    const cached =
+      (typeof sessionStorage !== 'undefined' && sessionStorage.getItem(key)) ||
+      (typeof localStorage !== 'undefined' && localStorage.getItem(key));
+    if (cached) {
+      const obj = JSON.parse(cached);
+      profileCache[uid] = obj;
+      return obj;
+    }
+  } catch {}
+
+  const usuarioSnap = await getDoc(doc(db, 'usuarios', uid));
+  const perfilSnap = await getDoc(doc(db, 'perfil', uid));
+  const mentSnap = await getDoc(doc(db, 'perfilMentorado', uid));
+  if (!usuarioSnap.exists()) return null;
+  const usuario = usuarioSnap.data() || {};
+  const perfil = perfilSnap.exists() ? perfilSnap.data() : {};
+  const perfilMentorado = mentSnap.exists() ? mentSnap.data() : {};
+
+  const data = {
+    uid,
+    perfil: usuario.perfil || perfil.perfil || 'usuario',
+    nome: usuario.nome || perfil.nome || '',
+    email: usuario.email || perfil.email || '',
+    isAdm: !!usuario.isAdm,
+    lojas: usuario.lojas || perfil.lojas || [],
+    ...perfil,
+    perfilMentorado,
+  };
+  try {
+    if (typeof sessionStorage !== 'undefined') {
+      sessionStorage.setItem(key, JSON.stringify(data));
+    }
+  } catch {}
+  profileCache[uid] = data;
+  return data;
+}
+
+export function clearUserProfileCache(uid) {
+  const key = `userProfile:${uid}`;
+  delete profileCache[uid];
+  try {
+    if (typeof sessionStorage !== 'undefined') {
+      sessionStorage.removeItem(key);
+    }
+  } catch {}
+}
+
+if (typeof window !== 'undefined') {
+  window.loadUserProfile = loadUserProfile;
+  window.clearUserProfileCache = clearUserProfileCache;
+}
 let wasLoggedIn = false;
 let authListenerRegistered = false;
 let explicitLogout = false;
